[Coder] Add logging, packaging, tests and CI
