+++ CHANGELOG.md
@@ -1,0 +1,11 @@
+# Changelog
+
+## 2025-07-24
+- Modularized code into collector package and added argparse CLI.
+- Collector now requires the `POLYGON_API_KEY` environment variable.
+
+## 2025-07-25
+- Added logging module with `--log-file` and `--log-level` CLI options.
+- Implemented incremental OHLCV downloading.
+- Added packaging metadata via `pyproject.toml`.
+- Created pytest tests for API functions and GitHub Actions CI.
