+++ NOTES.md
@@ -2,3 +2,15 @@
 - 2025-07-22 22:12 UTC: Created initial design notes under `design/` and populated `TASKS.md` with three starter tasks for securing API keys, adding tests, and introducing a configuration module.
 - 2025-07-22 22:35 UTC: Added architecture overview document and expanded TASKS.md with modularization tasks.
 - 2025-07-22 23:05 UTC: Documented next steps in `design/plan-2025-07-23.md` and extended task list with logging and CI tasks.
+
+- 2025-07-22 23:39 UTC: Added plan-2025-07-24.md with packaging objectives and outlined incremental extraction steps.
+- 2025-07-23 00:03 UTC: Planned next steps including logging module, tests, CI, packaging, and README updates. See plan-2025-07-25.md.
+
+# Coder
+- 2025-07-22 23:50 UTC: Modularized collector into package and added argparse CLI.
+- 2025-07-22 23:45 UTC: Removed fallback API key; script now requires POLYGON_API_KEY.
+- 2025-07-23 00:30 UTC: Added logging, incremental OHLCV, packaging metadata and basic tests.
+
+# Tester
+- 2025-07-23 00:28 UTC: Created pytest tests for API functions and verified they pass.
+
