+++ collector/__init__.py
@@ -1,0 +1,3 @@
+"""Collector package modules."""
+
+from .logging_utils import setup_logging
