+++ design/plan-2025-07-24.md
@@ -1,0 +1,13 @@
+# Planning Notes - 2025-07-24
+
+## Objectives
+- Kick off implementation of existing tasks starting with API key security and modularization.
+- Introduce packaging metadata so the collector can be installed as a CLI tool.
+
+## Proposed Tasks
+1. Add `pyproject.toml` with basic project metadata and entry point for `collector.main`.
+2. Document environment variable configuration for `POLYGON_API_KEY` and logging path.
+3. Plan incremental extraction of functions into the `collector` package.
+4. Ensure tests and CI run for every module as it is created.
+
+These notes set the direction for the next development cycle.
