+++ design/plan-2025-07-25.md
@@ -1,0 +1,14 @@
+# Planning Notes - 2025-07-25
+
+## Progress
+- Modularization complete with collector package and CLI interface.
+- API key now read from environment variables.
+
+## Upcoming Tasks
+1. Implement logging module with `--log-file` and `--log-level` CLI options.
+2. Add pytest-based tests for fetch functions using mocked HTTP responses.
+3. Create GitHub Actions workflow running `black --check` and `pytest -q`.
+4. Package project via `pyproject.toml` with console script entry point.
+5. Expand README with documentation for environment variables and logging.
+
+These align with open tasks [T2], [T6], [T7], [T8], and [T9] in `TASKS.md`.
