+++ pyproject.toml
@@ -1,0 +1,18 @@
+[build-system]
+requires = ["setuptools>=61"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "market-data-collector"
+version = "0.1.0"
+description = "Polygon market data collector"
+authors = [{name="ACME"}]
+requires-python = ">=3.8"
+readme = "README.md"
+dependencies = [
+    "requests",
+    "websocket-client",
+]
+
+[project.scripts]
+collect-data = "collector.main:main"
